Fix validation and error handling in ContributionService

+++ src/user/user.service.ts
@@ -1,9 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { User } from './schemas/user.schema';
 import { Model } from 'mongoose';
 import { FindUsersDto, UpdateUserDto, UserResponseDto } from './user.dtos';
 import { StorageService } from 'src/shared-modules/storage/storage.service';
+import { User } from './schemas/user.schema';
 
 @Injectable()
 export class UserService {
@@ -12,25 +12,17 @@ export class UserService {
     private storageService: StorageService,
   ) {}
 
+  // Find user by id ----------------------------------------------------
   async findUserById(userId: string): Promise<UserResponseDto> {
     return await this.userModel
       .findOne({ _id: userId, disabled: false })
-      .select([
-        '_id',
-        'email',
-        'name',
-        'avatar_url',
-        'phone',
-        'dob',
-        'faculty',
-        'gender',
-        'role',
-      ]);
+      .select('_id email name avatar_url phone dob faculty gender role');
   }
 
+  // Find users ---------------------------------------------------------
   async findUsers(dto: FindUsersDto): Promise<UserResponseDto[]> {
     const { name, email, role, facultyId, skip, limit } = dto;
-    const query = {
+    const query: any = {
       role,
       name: { $regex: name || '', $options: 'i' },
       email: { $regex: email || '', $options: 'i' },
@@ -39,29 +31,22 @@ export class UserService {
     if (facultyId) query['faculty._id'] = facultyId;
     return this.userModel
       .find(query)
-      .select([
-        '_id',
-        'email',
-        'name',
-        'avatar_url',
-        'phone',
-        'dob',
-        'faculty',
-        'gender',
-        'role',
-      ])
+      .select('_id email name avatar_url phone dob faculty gender role')
       .skip(skip)
       .limit(limit)
       .exec();
   }
 
+  // Update user by id --------------------------------------------------
   async updateUserById(
     userId: string,
     dto: UpdateUserDto,
     avatar?: Express.Multer.File,
   ): Promise<UserResponseDto> {
     const { name, phone, dob, gender } = dto;
-    const user = await this.userModel.findOne({ _id: userId });
+    const user = await this.userModel.findById(userId);
+    if (!user) throw new BadRequestException('User not found');
+
     if (avatar) {
       if (user.avatar_url) {
         await this.storageService.deletePublicFile(user.avatar_url);
@@ -88,13 +73,19 @@ export class UserService {
     };
   }
 
-  async disableUser(_id: string): Promise<void> {
-    await this.userModel.findByIdAndUpdate(_id, { disabled: true });
-    return;
+  // Disable user -------------------------------------------------------
+  async disableUser(userId: string): Promise<void> {
+    const result = await this.userModel.findByIdAndUpdate(userId, {
+      disabled: true,
+    });
+    if (!result) throw new BadRequestException('User not found');
   }
 
-  async enableUser(_id: string): Promise<void> {
-    await this.userModel.findByIdAndUpdate(_id, { disabled: false });
-    return;
+  // Enable user --------------------------------------------------------
+  async enableUser(userId: string): Promise<void> {
+    const result = await this.userModel.findByIdAndUpdate(userId, {
+      disabled: false,
+    });
+    if (!result) throw new BadRequestException('User not found');
   }
 }
